requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
apscheduler>=3.10.0
yfinance>=0.2.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
